+++ src/rwa/components/table/types/form.ts
@@ -65,8 +65,7 @@ export type GroupTransactionFormInputs = {
     fees?: TransactionFeeInput[] | null;
     cashAmount?: number | null;
     fixedIncomeAmount?: number | null;
-    unitPrice: number | null;
-    cashBalanceChange: number | null;
+    serviceProviderFeeTypeId?: string | null;
 };
 
 export type AccountFormInputs = {